Migrate button listener to button interaction
