+++ src/discord/interactions/buttons/types.ts
@@ -1,0 +1,6 @@
+import { ButtonInteraction } from 'discord.js';
+
+export interface Button {
+  buttons: string[]; // Array of button custom IDs or prefixes to listen for
+  onButtonInteraction: (interaction: ButtonInteraction) => Promise<void>; // The handler function
+}
